requirements.txt
streamlit
torch
torchvision
transformers
pillow
opencv-python
numpy
groq
python-dotenv
